update all packages and fixed problem with webgrease

+++ MobilePhonesOntology/ViewModels/StatsViewModel.cs
@@ -6,7 +6,7 @@ namespace MobilePhonesOntology.ViewModels
     {
         public int NumberOfBrandAndModels { get; set; }
         public int NumberOfPhones { get; set; }
-        public int NumbersOfRelations { get; set; }
+        public long NumbersOfRelations { get; set; }
         public int NumbersOfUniqueRelations { get; set; }
         public IEnumerable<string> NamesOfRelations { get; set; }
         public int NumbersOfUniqueProperties { get; set; }
